# Source/onespace-appdiscovery-service/src/main/resources/swagger/AppDiscoveryFramework.yaml
openapi: 3.0.3
info:
  title: OneSpace Platform APIs
  description: |-
    __OneSpace Platform APIs__ enables OneSpace Portal to register, list & approve Philips  serviceability applications.
  version: 1.0.0
servers:
  - url: https://apps.onespace-{region}-{env}.com/api
...
paths:
...
  /Notification:
    post:
      tags:
        - Notification APIs
      summary: Registers an notification with the given payload
      description: |
        A OAuth2.0 Bearer token of a subject to execute this operation.
      requestBody:
        description: Notification details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notification'
        required: true
      responses:
        '201':
          description: Notification triggered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
        '504':
          $ref: '#/components/responses/504'
    get:
      tags:
        - Notification APIs
      summary: Lists notifications entitled for the logged in user
      description: |
        A OAuth2.0 Bearer token of a subject to execute this operation.
      responses:
        '200':
          description: Notification retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationList'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
        '504':
          $ref: '#/components/responses/504'
  /Notification/$Read:
    post:
      tags:
        - Notification APIs
      summary: Marks the notifications as READ for the given set of noitification identifiers
      description: |
        A OAuth2.0 Bearer token of a subject to execute this operation.

      requestBody:
        description: Notification details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadNotificationList'
        required: true
      responses:
        '201':
          description: Notification read successfully
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
        '504':
          $ref: '#/components/responses/504'
  /Action:
    post:
      tags:
        - Action APIs
      summary: To register an action
      description: A session cookie with EPO_APP.LIST permission is required to execute this operation.
      requestBody:
        description: Action details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Action'
        required: true
      responses:
        '201':
          description: Action created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionID'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
        '504':
          $ref: '#/components/responses/504'
    get:
      tags:
        - Action APIs
      summary: To retrieve an action
      description: A session cookie with EPO_APP.LIST permission is required to execute this operation.
      responses:
        '201':
          description: Action retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionList'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
        '504':
          $ref: '#/components/responses/504'
    patch:
      tags:
        - Action APIs
      summary: Updates the action status
      description: |
        A Session Cookie with 
        __EPO_APP.UPDATE__ permission is required to execute this operation.

      requestBody:
        description: Action details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionPatchRequest'
        required: true
      responses:
        '200':
          description: Records are successfully updated.
          content:
            application/json:
              schema:
                type: string
                example: Records are successfully updated

        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '405':
          $ref: '#/components/responses/405'
        '406':
          $ref: '#/components/responses/406'
        '409':
          $ref: '#/components/responses/409'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
        '504':
          $ref: '#/components/responses/504'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          description: Error Description message
          type: string
          example: Error while processing Request
        code:
          description: Error code
          type: string
          example: error_code
        source:
          description: Error Source
          type: string
          example: application_discovery_service
        category:
          description: Error Category
          type: string
          example: error_category
    BadRequestErrorResponse:
      type: object
      properties:
        message:
          description: Bad Request Error Description
          type: string
          example: Invalid input data provided.
        code:
          description: Error code
          type: string
          example: bad_request
        source:
          description: Error Source
          type: string
          example: application_discovery_service
        category:
          description: Error Category
          type: string
          example: client error
    BadRequestError:
      type: object
      properties:
        issue:
          description: Error for Bad Request scenerios
          type: array
          items:
            $ref: "#/components/schemas/BadRequestErrorResponse"
    UnAuthorizedErrorResponse:
      type: object
      properties:
        message:
          description: UnAuthorized Operation Error Description
          type: string
          example: UnAuthorized Access
        code:
          description: Error code
          type: string
          example: invalid_token
        source:
          description: Error Source
          type: string
          example: application_discovery_service
        category:
          description: Error Category
          type: string
          example: invalid_token
    UnAuthorizedRequestError:
      type: object
      properties:
        issue:
          description: Error for Bad Request scenerios
          type: array
          items:
            $ref: "#/components/schemas/UnAuthorizedErrorResponse"
...
    ResourceNotFoundError:
      type: object
      properties:
        issue:
          description: Error for Bad Request scenerios
          type: array
          items:
            $ref: "#/components/schemas/ResourceNotFoundErrorResponse"
...
    ConflictError:
      type: object
      properties:
        issue:
          description: Error for Bad Request scenerios
          type: array
          items:
            $ref: "#/components/schemas/ConflictErrorResponse"
...
    InternalServerError:
      type: object
      properties:
        issue:
          description: Error for Bad Request scenerios
          type: array
          items:
            $ref: "#/components/schemas/InternalServerErrorResponse"
    Error:
      type: object
      properties:
        issue:
          description: A detailed human-readable message
          type: array
          items:
            $ref: "#/components/schemas/ErrorResponse"
...
    Notification:
      type: object
      properties:
        title:
          description: Notification title
          type: string
          example: Machine utilization alert
        message:
          description: Notification message
          type: string
          example: High Utilization Alert for CT Scanner 1
        category:
          description: Notification category
          type: string
          example: Informational, Actionable, Reminder, Alert, Confirmation, Promotional, Feedback, Error/Warning, Security, Maintenance
        source:
          type: object
          properties:
            id:
              description: Source application id, unique ID issued by OneSpace that represents a source. Such as an application
              type: string
              example: "550e8400-e29b-41d4-a716-446655440000"
        recipients:
          type: array
          items:
            $ref: "#/components/schemas/Recipient"
        createdAt:
          type: string
          format: date-time
          description: DateTime In UTC when this notification was generated
          example: 2024-09-18T13:09:05
        expiry:
          type: string
          format: date-time
          description: Date & Time after which notification will no longer be sent to the consumer
          example: 2024-09-25T13:09:05
    NotificationResponse:
      type: object
      properties:
        id:
          description: Notification id, unique ID issued by OneSpace that represents a notification.
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          description: Notification title
          type: string
          example: Machine utilization alert
        message:
          description: Notification message
          type: string
          example: High Utilization Alert for CT Scanner 1
        category:
          description: Notification category
          type: string
          example: Informational, Actionable, Reminder, Alert, Confirmation, Promotional, Feedback, Error/Warning, Security, Maintenance
        source:
          type: object
          properties:
            id:
              description: Source application id, unique ID issued by OneSpace that represents a source. Such as an application
              type: string
              example: "550e8400-e29b-41d4-a716-446655440000"
            name:
              description: Source application name
              type: string
              example: "Application Name"
        recipients:
          type: array
          items:
            $ref: "#/components/schemas/Recipient"
        createdAt:
          type: string
          format: date-time
          description: DateTime In UTC when this notification was generated
          example: 2024-09-18T13:09:05
        expiry:
          type: string
          format: date-time
          description: Date & Time after which notification will no longer be sent to the consumer
          example: 2024-09-25T13:09:05
        read:
          type: boolean
          example: false
    NotificationList:
      type: array
      items:
        $ref: "#/components/schemas/NotificationResponse"
    NotificationRead:
      type: string
      description: Notification guid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Recipient:
      type: string
      description: Recipient guid
      example: "550e8400-e29b-41d4-a716-446655440000"
    ReadNotificationList:
      type: array
      items:
        $ref: "#/components/schemas/NotificationRead"
...
    Action:
      type: object
      properties:
        initiator:
          description: The logged user
          type: string
          format: uuid
          example: 1f5f4f59-e251-4f19-a6bf-1528eeb8bb2f
        potentialOwner:
          description: The user to whom the action is assigned
          type: array
          items:
            type: string
          example: [ "1f76732b-861b-4997-b1f4-788eab3c0cba",
                     "7583f5f4-ea78-442a-99f4-0caddd45fb19"]
        status:
          description: The status of the action. By default, it is todo.
          type: string
          enum: [todo, done, expired]
          example: todo
        dateTime:
          type: string
          format: date-time
          description: Date and time when this action was created.
          example: 2024-09-20T13:09:05
        type:
          type: string
          enum: [informational, reminder, alert, feedback]
          example: alert
        title:
          description: The title for the action.
          type: string
          example: Reduce temperature in control room
        notify:
          description: This represents the notification flag. If set to true, the creation of an action generates a notification. By default it is false.
          type: boolean
          example: true
        message:
          description: Message that describes the action.
          type: string
          example: The temperature in the control room is outside a threshold. Measured value 60 degrees centigrade
        dueDateTime:
          type: string
          format: date-time
          description: The date & time in UTC by when the specific action should be completed.
          example: 2024-09-30T13:09:05
        expiryDateTime:
          type: string
          format: date-time
          description: The date & time in UTC by when the specific action is no longer relevant.
          example: 2024-10-05T13:09:05
        metadata:
          description: Additional information as key-value pairs for extensibility.
          type: object
          additionalProperties:
            type: string
          example:
            location: L1234
        relatedObject:
          description: The reference to the object (tile, widget, data grid, etc.) associated with the action
          type: string
          example: "https://www.google.com/"
        source:
          type: object
          properties:
            id:
              description: Unique ID generated by OneSpace that represents a source. Eg. application
              type: string
              example: "550e8400-e29b-41d4-a716-446655440000"
    ActionList:
      type: array
      items:
        $ref: "#/components/schemas/ActionResponse"
    ActionResponse:
      type: object
      properties:
        actionId:
          description: Unique Id that represents an action.
          type: string
          example: "67b5d981-1134-4600-8047-1a914b51391f"
        initiator:
          description: The logged user
          type: string
          format: uuid
          example: 1f5f4f59-e251-4f19-a6bf-1528eeb8bb2f
        potentialOwner:
          description: The user to whom the action is assigned
          type: array
          items:
            type: string
          example: [ "1f76732b-861b-4997-b1f4-788eab3c0cba",
                     "7583f5f4-ea78-442a-99f4-0caddd45fb19"]
        status:
          description: The status of the action.
          type: string
          enum: [todo, done, expired]
          example: done
        dateTime:
          type: string
          format: date-time
          description: Date and time when this action was created.
          example: 2024-09-20T13:09:05
        type:
          type: string
          enum: [informational, reminder, alert, feedback]
          example: alert
        title:
          description: The title for the action.
          type: string
          example: Reduce temperature in control room
        notify:
          description: This represents the notification flag. If set to true, the creation of an action generates a notification. By default it is false.
          type: boolean
          example: true
        message:
          description: Message that describes the action.
          type: string
          example: The temperature in the control room is outside a threshold. Measured value 60 degrees centigrade
        dueDateTime:
          type: string
          format: date-time
          description: The date & time in UTC by when the specific action should be completed.
          example: 2024-09-30T13:09:05
        expiryDateTime:
          type: string
          format: date-time
          description: The date & time in UTC by when the specific action is no longer relevant.
          example: 2024-10-05T13:09:05
        metadata:
          description: Additional information as key-value pairs for extensibility.
          type: object
          additionalProperties:
            type: string
          example:
            location: L1234
        relatedObject:
          description: The reference to the object (tile, widget, data grid, etc.) associated with the action
          type: string
          example: "https://www.google.com/"
        source:
          type: object
          properties:
            id:
              description: Unique ID generated by OneSpace that represents a source. Eg. application
              type: string
              example: "0d9972a1-5715-4f88-a3b5-e00826ae81b1"
            name:
              description: Source application name
              type: string
              example: "Customer Service Portal"
    ActionID:
      type: object
      properties:
        id:
          description: Unique Identifier that represents an action
          type: string
          format: uuid
          example: 120d45fc-314a-4d87-831d-d69f253263f6
    ActionPatchRequest:
      type: array
      items:
        $ref: "#/components/schemas/ActionUpdate"
    ActionUpdate:
      type: object
      properties:
        actionId:
          description: Unique Id that represents an action.
          type: string
          example: "67b5d981-1134-4600-8047-1a914b51391f"
        title:
          description: The title for the action.
          type: string
          example: Reduce temperature in control room
        status:
          description: The status of the action.
          type: string
          enum: [todo, done, expired]
          example: done
        initiator:
          description: The logged user
          type: string
          format: uuid
          example: 1f5f4f59-e251-4f19-a6bf-1528eeb8bb2f
        potentialOwner:
          description: The user to whom the action is assigned
          type: array
          items:
            type: string
          example: [ "7583f5f4-ea78-442a-99f4-0caddd45fb19"]
        completedAtDateTime:
          type: string
          format: date-time
          description: The date & time in UTC at which the action is completed.
          example: 2024-10-05T13:09:05
        source:
          type: object
          properties:
            id:
              description: Unique ID generated by OneSpace that represents a source. Eg. application
              type: string
              example: "0d9972a1-5715-4f88-a3b5-e00826ae81b1"
...
  responses:
...
    400:
      description: |
        **Bad Request**. Request is unparsable, syntactically incorrect, or violates 
        schema.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestError'
    401:
      description: |
        **Unauthorized**. Authorization failure due to missing or invalid session cookie.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/UnAuthorizedRequestError'
    403:
      description: |
        **Forbidden**. Operation is not permitted based on the supplied authorization.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/Error'
    404:
      description: |
        **Not Found**. Specified resource does not exist.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/ResourceNotFoundError'
    405:
      description: |
        **Method Not Allowed**. The method specified for the request is not allowed.
    406:
      description: |
        **Not Acceptable**. The media type is invalid.
    409:
      description: |
        **Conflict**. Request server refused to create or update a resource that already exists or has conflicting information.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/ConflictError'
    413:
      description: |
        **Request Entity Too Large**. The server is refusing to process a request because the 
        request entity is larger than the server is willing or able to process
    415:
      description: |
        **Unsupported Media Type**. The media type in the request is not supported.
    429:
      description: |
        **Too Many Requests**. The request quota has been filled. The server is throttling requests to prevent overload. 
    500:
      description: |
        **Internal Server Error**. The server encountered an unexpected condition which prevented 
        it from fulfilling the request.
      content:
        'application/json':
          schema:
            $ref: '#/components/schemas/InternalServerError'
    502:
      description: |
        **Bad Gateway**. The server received an invalid response from the upstream server.
    503:
      description: |
        **Service Unavailable**. The server is currently unable to handle the request due to a 
        temporary overloading or maintenance of the server.
    504:
      description: |
        **Gateway Timeout**. The server did not receive a timely response from the upstream 
        server specified it needed to access in attempting to complete the request 
...
  securitySchemes:
    onespaceCookieAuth:
      type: apiKey
      in: cookie
      name: onespacecookie
    bearerAuth:
      type: http
      scheme: bearer
security:
  - onespaceCookieAuth: []
  - bearerAuth: []